requirements.txt
# Core FastAPI stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client & WebSockets
httpx==0.25.2
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# Blockchain
web3==6.12.0
solana==0.32.0

# Async utilities
asyncio-mqtt==0.13.0
tenacity==8.2.3

# Scheduling
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0

# Development
flake8==6.1.0
flake8-bugbear==23.12.2
flake8-comprehensions==3.14.0
pep8-naming==0.13.3

# Decimal precision
decimal==1.70